+++ src/main/java/com/codeit/mission/deokhugam/book/dto/request/BookCreateRequest.java
@@ -3,6 +3,7 @@ package com.codeit.mission.deokhugam.book.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
+import java.time.LocalDate;
 
 public record BookCreateRequest(
     //제목
@@ -23,7 +24,7 @@ public record BookCreateRequest(
 
     //출판일
     @NotNull(message = "출판 일자를 적어주세요.")
-    Instant publishedDate,
+    LocalDate publishedDate,
 
     //ISBN 값
     String isbn
