+++ src/main/java/com/codeit/mission/deokhugam/book/dto/response/BookDto.java
@@ -1,6 +1,7 @@
 package com.codeit.mission.deokhugam.book.dto.response;
 
 import java.time.Instant;
+import java.time.LocalDate;
 import java.util.UUID;
 
 public record BookDto(
@@ -9,7 +10,7 @@ public record BookDto(
     String author,
     String description,
     String publisher,
-    Instant publishedDate,
+    LocalDate publishedDate,
     String isbn,
     String thumbnailUrl,
     int reviewCount,
