+++ src/main/java/com/codeit/mission/deokhugam/book/mapper/BookDtoMapper.java
@@ -1,6 +1,6 @@
 package com.codeit.mission.deokhugam.book.mapper;
 
-import com.codeit.mission.deokhugam.book.dto.BookDto;
+import com.codeit.mission.deokhugam.book.dto.response.BookDto;
 import com.codeit.mission.deokhugam.book.entity.Book;
 import org.mapstruct.Builder;
 import org.mapstruct.Mapper;
