+++ src/main/java/com/codeit/mission/deokhugam/book/repository/BookRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.codeit.mission.deokhugam.book.repository;
 
-import com.codeit.mission.deokhugam.book.dto.BookSearchConditionDto;
+import com.codeit.mission.deokhugam.book.dto.request.BookSearchConditionDto;
 import com.codeit.mission.deokhugam.book.entity.Book;
 import java.util.List;
